WIP: Statsforecast implementation
Based on discussion #198

Initial idea on how to implement Nixtla's Statsforecast into Functime,
it should be followed by NeuralForecast, however, not sure if all
parameters are exactly the same, hence, just wanted to give a simple
SF a try.

NOTE: This is a draft for now, so feedback is more than welcome.